SWIG: improve tool detection, and update SWIG tests to run on Windows.
The tool detection is improved by checking for swig in
env['SWIG'], where it is commonly set, as well as env['ENV']['PATH'].

The tests mostly didn't work on Windows.  I updated them all.
Mostly to build 32-bit extensions when using 32-bit python on
Windows, and use .pyd as the python extension on Windows.