Use a lock file when initializing the cache
There is a documented race condition when initializing the cache.
The race condition exists both between processes and threads within one
process, the latter is undocumented but more likely. Add a lock file so
that only one thread in one process at a time is allowed to initialize
a SCons cache.

Note that the implementation should work also for multiple computers
initializing a cache directory on a network drive, although that has not
been tested.

Signed-off-by: Fredrik Medley <fredrik.medley@autoliv.com>