First attempt to fix fortran issue with submodules.
I noticed that the scanner was already set up to ignore module followed
by the keyword procedure (case insensitive). Here, I then modify the
regex used for that to also ignore module followed by the keywords
subroutine or function. However, this does not fix the original problem
and I was able to confirm that SCons fails before the scanner is called
as printing out the value of the variable defmodules (not included in
the commit) is not printed. By changing the test files to something that
does not fail, I was able to confirm that the printout does happen in
that case and that module followed by subroutine and function are
correctly ignored.