Add error-on-warning control
Experimental: add --werror to allow specified warnings to be treated
as errors (exceptions).  Syntax is the same as --warn.  Warnings
have defaults built-in based on subclassing; scons maintains a list
of tuples as an override to hold explicit requests to warn or not
warn, populated from command line, SetOption, and a built-in table.
The list is looked up using a first-find basis, so it is sufficient to
insert a new value at the front.  The override list is not changed to
add warning-errors, but a new flag value (2) is added to mean raise
exception - there was already logic to raise the exception based on
the value of a global flag.

This is not ready for merge, it's more a proof of concept to generate
some discussion. The logic needs thinking through further.  The fact
that warnings are classes which may be heirarchical suggests that maybe
more examination should done on where exceptions are/are not generated.
There are no tests, and no documentation.  The messages are not great.
Here's an example where the message from the warning ends up being
deceptive:

  scons: *** ("Ignoring missing SConscript 'tizen/SConscript'",)

that is, it quits with a message which says it's ignoring a particular
problem, so it isn't ignoring it at all.

Signed-off-by: Mats Wichmann <mats@linux.com>