Update java version support
Although the newer versions are marketed as simple versions, scons will
see them as dotted versions: 9.0, 10.0.  Add some comments to this in
test code - the RMI test skip ought to work as long as they stay dotted
versions (since 10.0 >= 1.8); that code does run the test for a simple
version like '6'.

Signed-off-by: Mats Wichmann <mats@linux.com>