Add ability for SConscript to fail on missing script
SConscript call now takes an optional must_exist flag, which defaults
to False for compatiility with current behavior.  If True, an exception
is raised if the file is missing.  To improve readability, the
decision is moved off to a new function rather than being inline in
_SConscript.

A global setting to control the overall behavior is also added.

A deprecation warning is added for the current behavior, which is
printed only once.

Signed-off-by: Mats Wichmann <mats@linux.com>