Fix py3 ResourceWarning in TestCmd
class TestCmd method read() uses a shortcut to return data from
a file, "return open(...).read()".  Python 3 warns this is a
resource leak because the file has no chance to be closed with
the open being in the return statement.  Split into two lines
and use a context manager (with statement).

Signed-off-by: Mats Wichmann <mats@linux.com>