Testing: python 3 fix for must_contain
TestCommon defines a method must_contain which checks for a file including
a given string. With Python 3, the test runs into some typing problems.
This could be fixed either by changing all the tests which call the
routine either omitting the mode argument (which then defaults to 'rb'),
or specifying a mode which includes 'b'; or by modifying must_contain to
align the types of the file data and the data to check for. This patch
uses the latter approach.

This is a test-only change, no run-time scons code is modified.

Signed-off-by: Mats Wichmann <mats@linux.com>