Fix rpm tests for newer rpmbuild (fixes #3164)
Two issues: (1) ARCHITECTURE, if supplied, actually needs to be set
in the specfile BuildArch: field (or passed on the command line).
(2) rpmbuild on Fedora builds all kind of debug stuff behind the
scenes (it makes packages not requested by the specfile!). This
causes scons rpm tests to fail, so force in a flag to turn off this
generation. That is done unconditionally at the moment by writing

  %global debug_package %{nil}\n

to the specfile.  If anyone eventually ends up needing to produce
these debug packages through scons, we can add a flag to allow
it by omitting the above setting.

Some doc and comment cleanup done, and one of the tests now builds
a noarch package to test actually supplying an ARCHITECTURE value
that is not the machine default. Also, one of the tools called by
rpm packaging did not supply the right suffix for the Tar builder,
presumably a copy-paste error in a place that is not stressed by
any test.

Fedora packaging now also generates a file containing the build id
and puts it into the package (another behind the scenes bit of magic -
this does not appear in our %files section but is still added to the
package). It is possible to turn this behavior off as well, but for the
tests affected, the choice was made instead to have them now check the
files in the generated package /contain/ the expected filenames rather
than /exactly-match/ the expected filenames.

Testing: on recent Fedora all the rpm-related tests now pass after six
failed previously; on a Centos 7.4 system which did not fail any tests,
the modified code still passes - did not introduce regressions. At
the moment do not believe any of the CI setup will cause rpm tests
to be run.

Signed-off-by: Mats Wichmann <mats@linux.com>