Quiet Py3.6+ ResourceWarnings on tests
When a test calls _subproc to set up a subprocess.Popen call, if the
arg vector contains the string 'devnull' os.devnull is opened and the
resulting file object is stored in the kw vector prior to passing it to
Popen.  Reigster a close() call with atexit in this case.  Eliminates a
lot of Python 3.6+ ResourceWarning messages - 336 on this machine.

Signed-off-by: Mats Wichmann <mats@linux.com>