Remove unintended "import io"
During develoment of PR#3177, an interation checked objects for being
a subclass of io.IOBase, which necessitated importing the io module.
Later the check was changed to see if the object simply had a close
method, in which case that is called. The import was not removed;
cleaning that up.

Signed-off-by: Mats Wichmann <mats@linux.com>