Provide alternatives to md5 signatures
Work in progress (see TODO)

Broadly, two changes

(1) in the Visual Studio area, use uuid (Python standard library)
to generate GUID identifiers instead of specifically using md5.
Simplifies things as uuid lib can provide already formatted strings
and that does not have to be done in place.
TODO: some of the wired-in GUIDs in tests need to change to what
is being produced now.

(2) For use by file hashes, fall back to sha1 if md5 is not available
(rare, but documented in Python documentation as being possible in a
"FIPS Compliant" build of Python, and actually encountered by an scons
user); while considering hashslib itself always present.  Like uuid,
this is Python standard library since well before the oldest supported
Python version (both added in 2.5).  Util.py now describes the hash
being used in Util.md5, this is used to distinguish expected values in
tests. Except for tests, nobody should need to look inside: Util.md5
will still evaluate True.

Security comment: the hashes are not used for any security purpose
whatsoever, only for file-change detection, which in turn is used only for
rebuild calculations, not for tampering detection. Thus the alternative
is sha1, which also is "not secure", but which is as fast (in recent
Py3 apparently even faster) as md5 hash computation. An additional
fallback, should anyone develop problems with sha1 is sha256.

Signed-off-by: Mats Wichmann <mats@linux.com>