small type-related cleanups
Two files in packaging:
ipk.py indexed off the result of running filter, but in Python 3
filter returns an iterable, not a list. Convert to a list first.

msi.py removes forbidden characters using a list comprehension,
but the result is a list, so when it then calls upper() on it
that's a method that does not exist on a list. Join it back into
a string. Found another place in the same file that also assumed
the list comprehension leaves a string, not a list, although it
doesn't then call a nonexistent method on it.

Signed-off-by: Mats Wichmann <mats@linux.com>