Do not use time.clock for Py3 in runtest
time.clock is deprecated since Python 3.3 and will be removed in
Python 3.8.  3.7 started issuing DeprecationWarning, which
fails about 16 tests in AppVeyor CI (and running manually),
since the warning appears in stderr stream and so real vs expected
does not match any longer.  Arguably the tests could be fixed to
do a different check (contains rather than exactly-equal), but
a change needs to be made anyway.

use time.perf_counter as the time function for Python 3.  This
works for Windows and non-Windows; however since this function
did not exist in 2.7 (added for 3.3), the Py2 case is left, which
selects between time.clock and time.time depending on OS.

Note this addresses runtest.py, but not code in bench/, which
need attention also.  Suggest that be a separate change since
it's not failing tests.

Signed-off-by: Mats Wichmann <mats@linux.com>