generalize the fix for #3164
Add a new flag, X_RPM_EXTRADEFS, which lets the user supply
any additional flag lines needed in the specfile. If the
tag is not present in the specfile, supply a value which
turns off debug package generation. To retain debug package
generation, supply the flag with a value of None, or with
a list that does not include the disable value.

So like the previous change the default is now not to
generate debug, but that default can be overridden, and it
is documented.