Improvements to tarball packager
Both gz and bzip2 now skip the test if tar is not found (previously they
would not test, but mark the test OK).

Since tar is now found on recent Windows 10, but the helper program
bzip2 is not, check for bzip2 as well.  This is actually a correct test
for other systems as well, they have just been very unlikely to not have
had bzip2 provisioned so it has not bitten us.

Note: on Windows, with cygwin installed, the test may well find bzip2
but c:\Windows\System32\tar.exe does not find it and the test fails
anyway. Work needed on this "finds too much" issue (not unique to
this test).

Signed-off-by: Mats Wichmann <mats@linux.com>