[WIP] try to improve debug=explain
When variant dir is in effect, some deceptive reporting can happen.
In some circumstances, a file (changed or unchanged) may be reported
as no longer a dependency, then added as a dependency, when neither
is true.

This phenomenon is mentioned in at least #2932, #2935, #2936.

There are several contributing factors for the problem; these changes
just attempt to let explain behave more reasonably, but are not the
"natural" answer, part of which seems too hard to fix, and part of
which might be something funky in the variant-dir calculations.

Along the way, implement the suggestion from issue #2932 -
don't exclude report for changed action even if other rebuild
reason was already found.

Signed-off-by: Mats Wichmann <mats@linux.com>