explain now checks by names first (for PR #3252)
Introduce a new "data structure" - a list where the keys are the pathname
and the values are a tuple of the node and the signature. This way we can
easily check by name, but also have the node available for the "changed"
check that needs to be made.

A few minor pep8 changes were snuck in - spaces added/removed and blanks
lines added.  This is nowhere near a comprehensive pep8 cleanup.

Signed-off-by: Mats Wichmann <mats@linux.com>