scons-local version has symlink problems
Earlier fix to running with scons-local when the directory is a
symlink (i.e. scons-local symlinked to scons-local-{versiontag} and
invoked as scons-local/scons.py) broke running scons when the
"executable" is a symlink. Restoring that behavior.

This is dodgy to test reliably on a developer system, as the
fallback mechansim to find a usable scons goes on to look for
a system-installed scons, so it looked like it was working. I
will add a test if I can figure out a good way.

Signed-off-by: Mats Wichmann <mats@linux.com>