Align way TeX and related tool names reported by tests
Two concrete changes:

DVIPS/DVIPS test ran only if dvips tool is found, but did not leave
any message if not. Print a skip message if tool not found, which
also lets us dedent the rest of the test for the case where we
continue due to having found it.

TEX/variant_dir test did not check for dvipdf tool, but then calls
the tool. Add a check so we get a skip message instead of a fail
(the fail had lots of lines but it hard to see actual reason).

Fix one test which failed for me due to not checking for its tool.
For the rest, align "Could not find" messages to quote the tool name,
which most of the tests do - so this is just stylistic, has no
functional effect. Also stylistic: use any/all for checking
multiple tools' existence.

Signed-off-by: Mats Wichmann <mats@linux.com>