Fix problem where API doc gen misses some files
In doc/SConscript, the function in bootstrap.py to parse a
manifest was called with an scons path (#src/engine) but bootstrap
does not run in an scons context, it is run via subprocess.Popen.
Preprocess the path. Along the way, change the function to open
the file itself instead of being passed a list already read
from the file, update the comment, and clean some whitespace issues.

Signed-off-by: Mats Wichmann <mats@linux.com>