Add textfile tool to defaults
PR #3242 added the Textfile and Substfile builders to the default builder
list (for issue #3147), but didn't finish the job: the textfile tool
needs to be added to the default list of tools as well.  This time
with a testcase that fails if the tool is not added. Minor doc tweak.

Signed-off-by: Mats Wichmann <mats@linux.com>