[PYPY3] [PY 3.8] CacheDir tests use context managers
Python 3.8 complains about unclosed files on nearly all of the
CacheDir tests. PyPy3 fails 8 of the 20 tests for the same reason,
though since it's based on an earlier version of Python (3.5 at the
moment) it does not display the resourcewarning messages.

Update sequences of open().write() which are the primary problem
area to use context managers for automatic closing, and for
consistency, the rest of the open/read/write stuff as well.
With this change, all these tests pass in PyPy3. Python 3.8 does
not, as it is spewing other warning messages which also end up in
the stderr, which is fatal to this set of tests, but it has
quieted the warnings from the CacheDir tests themselves.

Signed-off-by: Mats Wichmann <mats@linux.com>