[PYPY] fix way exit status is retrieved in main
When scons exits, it wants to take the opportunity to print any
diagnostics and statistics that may have been requested, so the
main routine traps the various ways it can quit.  If code somewhere
calls sys.exit(), that generates a SystemExit exception.
The handling of that has not been quite correct - it simply takes
the exception instance, saves it, and later quits with sys.exit(saved).
This seemingly works fine for all other tested versions of Python,
but has interesting side effects with PyPy3.

Per the Python documentation:

If the value is an integer, it specifies the system exit status
(passed to C’s exit() function); if it is None, the exit status
is zero; if it has another type (such as a string), the object’s
value is printed and the exit status is one.

And in fact, PyPy3 does this: if the original call to sys.exit
took a value of 2, then the SystemExit exception triggers with
a class instance which does have an exit code of 2, which turns
up if you take the string repr of the instance, but when passed
to the final sys.exit, as it's not an integer it *prints* the 2,
and returns with an exit code of 1, just as in the documentation snip.

Not really sure if PyPy3 is wrong here, or the other Pythons are
letting something slide, but it's an easy fix: save off the code
stored in the exception instance for later use as the exit code,
instead of using the instance itself for that.

Signed-off-by: Mats Wichmann <mats@linux.com>