Fix issue #3135.
This fixes issue #3135 regarding the issue with using type bound
procedures in Fortran submodules. The fix consists of changing the
regex used in the scanner and the emitter to ignore lines starting
with:

module subroutine

and

module function

as these are used to define type bound procedures instead of modules
named 'subroutine' or 'function'.  The regex is case insensitive.