Clean up some file opens, regex strings
Most recent Python (3.8 alpha) spews warnings aplenty about two
subjects: unclosed files and strings which look like they have embedded
escapes that Python does not recognize.  The latter are usually
regexes, and it provides a reminder that regular expressions should
normally be specified as raw strings, so Python does not attempt to
interpret them.  Irritating is that even docstrings are flagged, it's
not obvious what the right answer is for a docstring which contains,
say, a Windows-style path with backslashes.

This converts a bunch of opens that are not closed into context
manager usage and regex patterns into raw strings.

This eliminate about 4000 warnings spewed by Py3.8 (9200 remain).

Signed-off-by: Mats Wichmann <mats@linux.com>