Runtest improvements for multi-job runs
Main fix: if duing a multi-job run (-j flag is present), collect the
test output in runtest and keep the test heading together with the test
output, so the information about a given test will not separated in
the output.  The test number is recorded in the test instance instead
of kept as a global.  Tests may appear out of order now, but since they
may be run "out of order" depending on how fast the various threads
complete the jobs they pull off the queue, that just reflects reality.

The help message is now the script's docstring, and help prints the
docstring along with some other info. This avoids having the options
described twice (and going out of sync) - once in the inital comment,
once in the help message.  A mention of the -j multi-job option is
added, was previously missing.

Also: one less worker than requested was created due to use of
range(1, n) (may have been intentional?). changed to number requested.

Also: the lock is taken from the RunTest instance, not from the global
lock variable.  This must have been intended as the lock is stored in
the instance on initialization. The lock was not really declared as
a global, it was just assigned in global scope.

Also: the test result reporting is now split off into its own function,
makes a bit cleaner separation (and facilitates possible changes to
the multi-job model in future).

Signed-off-by: Mats Wichmann <mats@linux.com>