Fix problems with jdk detection
The java tool common routine finds a jdk by doing a filesystem glob.
This had a problem on windows in the case a specific version is requested,
because the format of name of the jdk directory has changed with JDK 9 -
there is a dash between jdk and the version string.  The glob which does
not attempt to match a version was general enough not to trip on this,
but with a version to match it would never match jdk-9 or higher.

The test harness then asks the found javac what version it is, and the
parsing of that did not work as expected once version numbers became
double-digit, as the regex was for a single digit followed by a dot.
The outcome is for 11.0.2 we get back '11' instead of '11.0'.  Change the
regex to match any number of digits followed by dot.

The Repository/RMIC.py change is to align with an earlier change to
Java/RMIC.py, but the may not be needed after the change to regex
just described.

Clean up some of the Java tool routines for consistency (there was
no functional change outside of JavaCommon.py)

Docstrings added or updated in several places.

Signed-off-by: Mats Wichmann <mats@linux.com>