[PR #3330] try to eliminate race in scons-time
runtest change to run requested number of jobs exposed a concurrency
problem on Appveyor CI (Windows) builds - which seems to be due to
scons-time using mktemp to create a temporary directory name, and then
later creating the directory.  In scons-time the timing window for
this is a bit longer than usual.  Try using mkdtemp instead.

Signed-off-by: Mats Wichmann <mats@linux.com>