[PR #3352] fix sider complaints: bare except
Sider in the CI system doesn't like the change which added a try/except
block for a version without a dot.  This was copied from a different
file, so to fix the complaint, four instances are actually updated to
"except ValueError" - I am assuming that this is the only thing that
could go wrong, a version string cannot be converted with int(foo),
which would throw ValueError.

Signed-off-by: Mats Wichmann <mats@linux.com>