Fixup some code triggering pylint errors.
Assorted fixups: exception types, redefined functions, globals, etc.

Some old code removed to resolve issues (hashlib is always present on
modern Pythons; no longer need the code for 2.5-and-earlier optparse).

cmp is not a builtin function in Py3, drop one (unused) use; replace one.

Fix another instance of renaming to SConsEnvironmentError.

TODO flagged some instances of doing a raise without argument but not
inside a try block - this is not considered legal, since raise
with no argument is for re-raising an exception, but I don't know
exactly how to resolve this in these cases. Also flagged an instance
of raising an int instead of an exception class. We can either leave
these as markers or update the PR.

Signed-off-by: Mats Wichmann <mats@linux.com>