Also ignore PURE and ELEMENTAL after MODULE in Scanner and Emitter.
Add subroutines that are declared pure and elemental to the test
of the Emitter. Note that in test_1.f90, the interface is repeated
in the submodule, whereas in test_2.f90 the interface is taken from
the module. Also note that the regex does not check whether
"module pure" or "module elemental" is actually followed by "subroutine" or
"function". This would be a syntax error and should trigger a compile time
error.