Some more lint-derived cleanups
Consistently use "not is" and "not in", many instances used
the form "not x is y" instead, which pylint objected to.
A couple of bare except clauses got a qualifier.
Files otherwise touched had trailing whitespace cleaned up as well.

These are all things that sider would complain about if a change
happened nearby, so this is pre-emptive.

Signed-off-by: Mats Wichmann <mats@linux.com>