test harness: fix TestCommon tests
The testing/harness/TestCommonTests.py unit tests do not currently
pass.  Note they are not run by the CI system, or in fact by
doing runtests -a, since they're in a directory that is not searched.
After these changes, there are no fails.

Also, the methods that used Google-style docbook markup are changed
to REsT-style markup.  Our current doc producer, epydoc, does not
understand the Google style, and we shouldn't mix styles; can convert
them all in bulk later if we switch to Sphinx as the production tool.