More travis-ci experiments
the procedure to get a pkg from a non-Ubuntu repository
is updated - it's gotten fussier recently. This wouldn't
matter for 14.04 and seems not to for 16.04, but better
matches the info on the current D site, plus adds the
extra fuss needed for 18.04 and later.

Versions bumped on ldc and swig.

Pkg install list fiddled a bit - it will probably be
wrong and need more fiddling.

Signed-off-by: Mats Wichmann <mats@linux.com>