test harness: fix TestCmd tests
This is the companion to PR #3382.  The testing/harness/TestCmdTests.py
unit tests do not currently pass. Note they are not run by the CI
system, or in fact by doing runtests -a, since they're in a directory
that is not searched.  After these changes, there are no fails. This is
a test-only change.

The method simple_diff, modeled on difflib functions, is converted to
a generator to match difflib, and now has a doctest as well.  This
means calls to it which aren't going to iterate needs to convert the
return with list(), but that just makes usage more consistent, since
the calls to difflib.context_diff and difflib.unified_diff already had
to do so.

Also, the methods that used Google-style docbook markup are changed to
REsT-style markup. Our current doc producer, epydoc, does not understand
the Google style, and we shouldn't mix styles; can convert them all in
bulk later if we switch to Sphinx as the production tool.

Signed-off-by: Mats Wichmann <mats@linux.com>