Prototype of a more aggressive parallel scanner
This is a work-in-progress changeset for a more aggressive parallel scanner.
The current parallel scanner wastes time waiting for thread pool results when
it could have been more aggressively continuing to gather more tasks to
dispatch. The idea here is that if we have dispatched the maximum number of
jobs and none are done yet, we should gather more tasks instead of doing a
blocking wait on the thread pool. In my testing, I have seen a 4-5% improvement
in some VMware builds.

One upside of this change is that it paves the way for us to be able to
implement remote caching because we now ask the cache for groups of tasks
instead of one-at-a-time.

At VMware we don't have a good mechanism for testing upstream SCons changes so
this is a blind application of a patch I wrote. It passes some tests but
may have some bugs compared to the version I applied internally. More work
needed here.