test: Fix SyntaxWarnings on escapes in Python strings
This is the remaining batch of SyntaxWarning cleanups in the test/
directory, as flagged by Python 3.8 (now tested with 3.8.0b4).  Most of
these are escapes in strings that are not recognized as Python-language
escapes up front, due to the embedded nature of testing - tests write a
string to a file which is actually Python (code or sconscript) which is
later to be executed.

Also changed the timestamp-fallback test to eliminate an imp module
DeprecationWarning: since in all supported Python versions haslib
is stdlib, drop that skip check; the md5 skip check can be a simple
import now.

My test runs with 3.8 show no SyntaxWarnings after this.

Signed-off-by: Mats Wichmann <mats@linux.com>