Stop converting to list where not needed
Python 3 returns a special object, which is iterable, rather than
a list when you ask for dictionary keys(), values(), items(). if
you then proceed to iterate over it it's being used as expected and
doesn't have to be forced to a list first.  This occurs a number of
places in this form:

  for k in list(something.keys()):

Also there are several places where the code loops over the result
of dict.keys() and then uses the key to index into the dictionary,
this can be replaced by:

  for k, v in something.items():

Signed-off-by: Mats Wichmann <mats@linux.com>