Actually say it's the MIT license   [ci skip]
A line "MIT License" is added to the license files, most
places do it this way.

For the benefit of Github, the top-level LICENSE file, which it
looks at but the packaging stuff does not, is filled-in. The
packaging fills in __COPYRIGHT__ from src/LICENSE.txt instead
(and LICENSE-local).

Add a skeletal CONTRIBUTING.md

Signed-off-by: Mats Wichmann <mats@linux.com>