Builder and Variable Subst doc updates  [ci skip]
Work on Builder Methods and Variable Substitution man sections - some
wording tweaks, do some formatting markup a little more consistently,
etc.

Tweak the wording for the Command builder, a description shared
between manpage and userguide.

Closes #3030: Environment.xml had many references to detail sections
elsewhere, like "see XXX section below," which don't work well -
this xml is generated into both the manpage and the user guide,
and the locality of saying "below" is incorrect for the latter since
those sections are only in the manpage. Essentially these are changed
to say "see the manpage section XXX". Adding clickable links is of
course also possible, but I recall there was some objection to
cross-document links, so I didn't do that.

Signed-off-by: Mats Wichmann <mats@linux.com>