Remove deprecated debug options
These options have been deprecated since 2007.  They were originally
announced to be disabled in SCons 2.0.0, but that didn't happen.

Left the deprecated-debug-options behavior is in, but the dictionary of
such options is now empty, and there's a new dict of removed options,
and presence in that dict raises an exception.

The four tests that were in test/Deprecated move to a new directory
test/Removed and are simplified just to make sure invocation errors
scons out. (git interprets most of these as remove/add for some reason)

These appear to have been already removed from docs, so no doc impact.

Signed-off-by: Mats Wichmann <mats@linux.com>