[PR 3462] convert json cache better way for py2
there was still a path through without fully converted data,
where env['ENV'] could contain unicode, so convert the cache hit instead.
This is a little slower but only for Py2 which is going away anyway.

added the vs2019 exec test (vs-14.2-exec) which is just a copy
of 14.0, 14.1 with the required version changed - this had
never been added.

Signed-off-by: Mats Wichmann <mats@linux.com>