Add support for scanning multiple entries in an action string
This change adds support for scanning multiple entries in an action string in
order to better support the following use cases:

1. A file is provided in an action string and should be taken as a dependency.
   For example, an action string "$PERL somefile.pl".
2. An action string actually has two actions separated by &&. For example,
   "cd <some_dir> && $ZIP <args>".

Adding support for #1 actually allows us to fix the test
IMPLICIT_COMMAND_DEPENDENCIES.py on Windows, which was previously treating a
Python file as executable even on Windows. This was causing tests to repeatedly
open the default handler of Python files, which if set to Visual Studio causes
DDE hangs. This test is fixed because now we can have the action string specify
python as the first command and still take an implicit dependency on the
script, which is now the second command.