[PR 3475] update docs/generated  [ci skip]
There's debated about whether the generated should be
checked in at all; without an answer to that, they've fallen
far enough out of sync that it becomes harder to use "git diff"
as a spot-check that nothing strange is happening due to any
given doc change.  This should resync them to current.

Signed-off-by: Mats Wichmann <mats@linux.com>